PolynomialSolutions

Generate polynomial solutions to certain PDEs given a polynomial source term.

Helmholtz equation

The function solve_helmholtz takes a Polynomial Q and returns the unique Polynomial P satisfying

\[ \Delta P + P = Q\]

Some solutions in 2D:

using PolynomialSolutions
for j in 0:4, i in 0:4
    Q = monomial(i,j)
    P = solve_helmholtz(Q)
    @show P, Q
end

(P, Q) = (1, 1)
(P, Q) = (x, x)
(P, Q) = (-2 + x², x²)
(P, Q) = (-6x + x³, x³)
(P, Q) = (24 - -12x² + x⁴, x⁴)
(P, Q) = (y, y)
(P, Q) = (xy, xy)
(P, Q) = (-2y + x²y, x²y)
(P, Q) = (-6xy + x³y, x³y)
(P, Q) = (24y - -12x²y + x⁴y, x⁴y)
(P, Q) = (-2 + y², y²)
(P, Q) = (-2x + xy², xy²)
(P, Q) = (-2y² - -2x² + x²y², x²y²)
(P, Q) = (-6xy² - -2x³ + x³y², x³y²)
(P, Q) = (24y² + 48x² - -12x²y² - -2x⁴ + x⁴y², x⁴y²)
(P, Q) = (-6y + y³, y³)
(P, Q) = (-6xy + xy³, xy³)
(P, Q) = (-2y³ - -6x²y + x²y³, x²y³)
(P, Q) = (-6xy³ - -6x³y + x³y³, x³y³)
(P, Q) = (24y³ + 144x²y - -12x²y³ - -6x⁴y + x⁴y³, x⁴y³)
(P, Q) = (24 - -12y² + y⁴, y⁴)
(P, Q) = (24x - -12xy² + xy⁴, xy⁴)
(P, Q) = (48y² - -2y⁴ + 24x² - -12x²y² + x²y⁴, x²y⁴)
(P, Q) = (144xy² - -6xy⁴ + 24x³ - -12x³y² + x³y⁴, x³y⁴)
(P, Q) = (24y⁴ + 288x²y² - -12x²y⁴ + 24x⁴ - -12x⁴y² + x⁴y⁴, x⁴y⁴)

And some solutions in 3D:

using PolynomialSolutions
for k in 0:4, j in 0:4, i in 0:4
    Q = monomial(i,j,k)
    P = solve_helmholtz(Q)
    @show P, Q
end

(P, Q) = (1, 1)
(P, Q) = (x, x)
(P, Q) = (-2 + x², x²)
(P, Q) = (-6x + x³, x³)
(P, Q) = (24 - -12x² + x⁴, x⁴)
(P, Q) = (y, y)
(P, Q) = (xy, xy)
(P, Q) = (-2y + x²y, x²y)
(P, Q) = (-6xy + x³y, x³y)
(P, Q) = (24y - -12x²y + x⁴y, x⁴y)
(P, Q) = (-2 + y², y²)
(P, Q) = (-2x + xy², xy²)
(P, Q) = (-2y² - -2x² + x²y², x²y²)
(P, Q) = (-6xy² - -2x³ + x³y², x³y²)
(P, Q) = (24y² + 48x² - -12x²y² - -2x⁴ + x⁴y², x⁴y²)
(P, Q) = (-6y + y³, y³)
(P, Q) = (-6xy + xy³, xy³)
(P, Q) = (-2y³ - -6x²y + x²y³, x²y³)
(P, Q) = (-6xy³ - -6x³y + x³y³, x³y³)
(P, Q) = (24y³ + 144x²y - -12x²y³ - -6x⁴y + x⁴y³, x⁴y³)
(P, Q) = (24 - -12y² + y⁴, y⁴)
(P, Q) = (24x - -12xy² + xy⁴, xy⁴)
(P, Q) = (48y² - -2y⁴ + 24x² - -12x²y² + x²y⁴, x²y⁴)
(P, Q) = (144xy² - -6xy⁴ + 24x³ - -12x³y² + x³y⁴, x³y⁴)
(P, Q) = (24y⁴ + 288x²y² - -12x²y⁴ + 24x⁴ - -12x⁴y² + x⁴y⁴, x⁴y⁴)
(P, Q) = (z, z)
(P, Q) = (xz, xz)
(P, Q) = (-2z + x²z, x²z)
(P, Q) = (-6xz + x³z, x³z)
(P, Q) = (24z - -12x²z + x⁴z, x⁴z)
(P, Q) = (yz, yz)
(P, Q) = (xyz, xyz)
(P, Q) = (-2yz + x²yz, x²yz)
(P, Q) = (-6xyz + x³yz, x³yz)
(P, Q) = (24yz - -12x²yz + x⁴yz, x⁴yz)
(P, Q) = (-2z + y²z, y²z)
(P, Q) = (-2xz + xy²z, xy²z)
(P, Q) = (-2y²z - -2x²z + x²y²z, x²y²z)
(P, Q) = (-6xy²z - -2x³z + x³y²z, x³y²z)
(P, Q) = (24y²z + 48x²z - -12x²y²z - -2x⁴z + x⁴y²z, x⁴y²z)
(P, Q) = (-6yz + y³z, y³z)
(P, Q) = (-6xyz + xy³z, xy³z)
(P, Q) = (-2y³z - -6x²yz + x²y³z, x²y³z)
(P, Q) = (-6xy³z - -6x³yz + x³y³z, x³y³z)
(P, Q) = (24y³z + 144x²yz - -12x²y³z - -6x⁴yz + x⁴y³z, x⁴y³z)
(P, Q) = (24z - -12y²z + y⁴z, y⁴z)
(P, Q) = (24xz - -12xy²z + xy⁴z, xy⁴z)
(P, Q) = (48y²z - -2y⁴z + 24x²z - -12x²y²z + x²y⁴z, x²y⁴z)
(P, Q) = (144xy²z - -6xy⁴z + 24x³z - -12x³y²z + x³y⁴z, x³y⁴z)
(P, Q) = (24y⁴z + 288x²y²z - -12x²y⁴z + 24x⁴z - -12x⁴y²z + x⁴y⁴z, x⁴y⁴z)
(P, Q) = (-2 + z², z²)
(P, Q) = (-2x + xz², xz²)
(P, Q) = (-2z² - -2x² + x²z², x²z²)
(P, Q) = (-6xz² - -2x³ + x³z², x³z²)
(P, Q) = (24z² + 48x² - -12x²z² - -2x⁴ + x⁴z², x⁴z²)
(P, Q) = (-2y + yz², yz²)
(P, Q) = (-2xy + xyz², xyz²)
(P, Q) = (-2yz² - -2x²y + x²yz², x²yz²)
(P, Q) = (-6xyz² - -2x³y + x³yz², x³yz²)
(P, Q) = (24yz² + 48x²y - -12x²yz² - -2x⁴y + x⁴yz², x⁴yz²)
(P, Q) = (-2z² - -2y² + y²z², y²z²)
(P, Q) = (-2xz² - -2xy² + xy²z², xy²z²)
(P, Q) = (-2y²z² - -2x²z² - -2x²y² + x²y²z², x²y²z²)
(P, Q) = (-6xy²z² - -2x³z² - -2x³y² + x³y²z², x³y²z²)
(P, Q) = (24y²z² + 48x²z² + 48x²y² - -12x²y²z² + 8x⁴ - -2x⁴z² - -2x⁴y² + x⁴y²z², x⁴y²z²)
(P, Q) = (-6yz² - -2y³ + y³z², y³z²)
(P, Q) = (-6xyz² - -2xy³ + xy³z², xy³z²)
(P, Q) = (-2y³z² - -6x²yz² - -2x²y³ + x²y³z², x²y³z²)
(P, Q) = (-6xy³z² - -6x³yz² - -2x³y³ + x³y³z², x³y³z²)
(P, Q) = (24y³z² + 144x²yz² + 48x²y³ - -12x²y³z² + 24x⁴y - -6x⁴yz² - -2x⁴y³ + x⁴y³z², x⁴y³z²)
(P, Q) = (24z² + 48y² - -12y²z² - -2y⁴ + y⁴z², y⁴z²)
(P, Q) = (24xz² + 48xy² - -12xy²z² - -2xy⁴ + xy⁴z², xy⁴z²)
(P, Q) = (48y²z² + 8y⁴ - -2y⁴z² + 24x²z² + 48x²y² - -12x²y²z² - -2x²y⁴ + x²y⁴z², x²y⁴z²)
(P, Q) = (144xy²z² + 24xy⁴ - -6xy⁴z² + 24x³z² + 48x³y² - -12x³y²z² - -2x³y⁴ + x³y⁴z², x³y⁴z²)
(P, Q) = (24y⁴z² + 288x²y²z² + 48x²y⁴ - -12x²y⁴z² + 24x⁴z² + 48x⁴y² - -12x⁴y²z² - -2x⁴y⁴ + x⁴y⁴z², x⁴y⁴z²)
(P, Q) = (-6z + z³, z³)
(P, Q) = (-6xz + xz³, xz³)
(P, Q) = (-2z³ - -6x²z + x²z³, x²z³)
(P, Q) = (-6xz³ - -6x³z + x³z³, x³z³)
(P, Q) = (24z³ + 144x²z - -12x²z³ - -6x⁴z + x⁴z³, x⁴z³)
(P, Q) = (-6yz + yz³, yz³)
(P, Q) = (-6xyz + xyz³, xyz³)
(P, Q) = (-2yz³ - -6x²yz + x²yz³, x²yz³)
(P, Q) = (-6xyz³ - -6x³yz + x³yz³, x³yz³)
(P, Q) = (24yz³ + 144x²yz - -12x²yz³ - -6x⁴yz + x⁴yz³, x⁴yz³)
(P, Q) = (-2z³ - -6y²z + y²z³, y²z³)
(P, Q) = (-2xz³ - -6xy²z + xy²z³, xy²z³)
(P, Q) = (-2y²z³ - -2x²z³ - -6x²y²z + x²y²z³, x²y²z³)
(P, Q) = (-6xy²z³ - -2x³z³ - -6x³y²z + x³y²z³, x³y²z³)
(P, Q) = (24y²z³ + 48x²z³ + 144x²y²z - -12x²y²z³ + 24x⁴z - -2x⁴z³ - -6x⁴y²z + x⁴y²z³, x⁴y²z³)
(P, Q) = (-6yz³ - -6y³z + y³z³, y³z³)
(P, Q) = (-6xyz³ - -6xy³z + xy³z³, xy³z³)
(P, Q) = (-2y³z³ - -6x²yz³ - -6x²y³z + x²y³z³, x²y³z³)
(P, Q) = (-6xy³z³ - -6x³yz³ - -6x³y³z + x³y³z³, x³y³z³)
(P, Q) = (24y³z³ + 144x²yz³ + 144x²y³z - -12x²y³z³ + 72x⁴yz - -6x⁴yz³ - -6x⁴y³z + x⁴y³z³, x⁴y³z³)
(P, Q) = (24z³ + 144y²z - -12y²z³ - -6y⁴z + y⁴z³, y⁴z³)
(P, Q) = (24xz³ + 144xy²z - -12xy²z³ - -6xy⁴z + xy⁴z³, xy⁴z³)
(P, Q) = (48y²z³ + 24y⁴z - -2y⁴z³ + 24x²z³ + 144x²y²z - -12x²y²z³ - -6x²y⁴z + x²y⁴z³, x²y⁴z³)
(P, Q) = (144xy²z³ + 72xy⁴z - -6xy⁴z³ + 24x³z³ + 144x³y²z - -12x³y²z³ - -6x³y⁴z + x³y⁴z³, x³y⁴z³)
(P, Q) = (24y⁴z³ + 288x²y²z³ + 144x²y⁴z - -12x²y⁴z³ + 24x⁴z³ + 144x⁴y²z - -12x⁴y²z³ - -6x⁴y⁴z + x⁴y⁴z³, x⁴y⁴z³)
(P, Q) = (24 - -12z² + z⁴, z⁴)
(P, Q) = (24x - -12xz² + xz⁴, xz⁴)
(P, Q) = (48z² - -2z⁴ + 24x² - -12x²z² + x²z⁴, x²z⁴)
(P, Q) = (144xz² - -6xz⁴ + 24x³ - -12x³z² + x³z⁴, x³z⁴)
(P, Q) = (24z⁴ + 288x²z² - -12x²z⁴ + 24x⁴ - -12x⁴z² + x⁴z⁴, x⁴z⁴)
(P, Q) = (24y - -12yz² + yz⁴, yz⁴)
(P, Q) = (24xy - -12xyz² + xyz⁴, xyz⁴)
(P, Q) = (48yz² - -2yz⁴ + 24x²y - -12x²yz² + x²yz⁴, x²yz⁴)
(P, Q) = (144xyz² - -6xyz⁴ + 24x³y - -12x³yz² + x³yz⁴, x³yz⁴)
(P, Q) = (24yz⁴ + 288x²yz² - -12x²yz⁴ + 24x⁴y - -12x⁴yz² + x⁴yz⁴, x⁴yz⁴)
(P, Q) = (48z² - -2z⁴ + 24y² - -12y²z² + y²z⁴, y²z⁴)
(P, Q) = (48xz² - -2xz⁴ + 24xy² - -12xy²z² + xy²z⁴, xy²z⁴)
(P, Q) = (8z⁴ + 48y²z² - -2y²z⁴ + 48x²z² - -2x²z⁴ + 24x²y² - -12x²y²z² + x²y²z⁴, x²y²z⁴)
(P, Q) = (24xz⁴ + 144xy²z² - -6xy²z⁴ + 48x³z² - -2x³z⁴ + 24x³y² - -12x³y²z² + x³y²z⁴, x³y²z⁴)
(P, Q) = (24y²z⁴ + 48x²z⁴ + 288x²y²z² - -12x²y²z⁴ + 48x⁴z² - -2x⁴z⁴ + 24x⁴y² - -12x⁴y²z² + x⁴y²z⁴, x⁴y²z⁴)
(P, Q) = (144yz² - -6yz⁴ + 24y³ - -12y³z² + y³z⁴, y³z⁴)
(P, Q) = (144xyz² - -6xyz⁴ + 24xy³ - -12xy³z² + xy³z⁴, xy³z⁴)
(P, Q) = (24yz⁴ + 48y³z² - -2y³z⁴ + 144x²yz² - -6x²yz⁴ + 24x²y³ - -12x²y³z² + x²y³z⁴, x²y³z⁴)
(P, Q) = (72xyz⁴ + 144xy³z² - -6xy³z⁴ + 144x³yz² - -6x³yz⁴ + 24x³y³ - -12x³y³z² + x³y³z⁴, x³y³z⁴)
(P, Q) = (24y³z⁴ + 144x²yz⁴ + 288x²y³z² - -12x²y³z⁴ + 144x⁴yz² - -6x⁴yz⁴ + 24x⁴y³ - -12x⁴y³z² + x⁴y³z⁴, x⁴y³z⁴)
(P, Q) = (24z⁴ + 288y²z² - -12y²z⁴ + 24y⁴ - -12y⁴z² + y⁴z⁴, y⁴z⁴)
(P, Q) = (24xz⁴ + 288xy²z² - -12xy²z⁴ + 24xy⁴ - -12xy⁴z² + xy⁴z⁴, xy⁴z⁴)
(P, Q) = (48y²z⁴ + 48y⁴z² - -2y⁴z⁴ + 24x²z⁴ + 288x²y²z² - -12x²y²z⁴ + 24x²y⁴ - -12x²y⁴z² + x²y⁴z⁴, x²y⁴z⁴)
(P, Q) = (144xy²z⁴ + 144xy⁴z² - -6xy⁴z⁴ + 24x³z⁴ + 288x³y²z² - -12x³y²z⁴ + 24x³y⁴ - -12x³y⁴z² + x³y⁴z⁴, x³y⁴z⁴)
(P, Q) = (24y⁴z⁴ + 288x²y²z⁴ + 288x²y⁴z² - -12x²y⁴z⁴ + 24x⁴z⁴ + 288x⁴y²z² - -12x⁴y²z⁴ + 24x⁴y⁴ - -12x⁴y⁴z² + x⁴y⁴z⁴, x⁴y⁴z⁴)

Laplace

Missing

TODO

Docstrings

PolynomialSolutions.Polynomial — Type

struct Polynomial{N,T}

A polynomial in N variables with coefficients of type T.

PolynomialSolutions.degree — Method

degree(p::Polynomial)

Return the (maximum) degree of the polynomial p.

PolynomialSolutions.solve_helmholtz — Method

solve_helmholtz(Q::Polynomial)

Return the unique polynomial P satisfying ΔP + P = Q.