PolynomialSolutions

Generate polynomial solutions to certain PDEs given a polynomial source term.

Helmholtz

The function solve_helmholtz takes a Polynomial Q and returns the unique Polynomial P satisfying

\[ \Delta P + P = Q\]

Some solutions in 2D:

using PolynomialSolutions
for j in 0:4, i in 0:4
    Q = monomial(i,j)
    P = solve_helmholtz(Q)
    @show P, Q
end

And some solutions in 3D:

using PolynomialSolutions
for k in 0:4, j in 0:4, i in 0:4
    Q = monomial(i,j,k)
    P = solve_helmholtz(Q)
    @show P, Q
end

Laplace

The function solve_laplace takes a Polynomial Q and returns a Polynomial P satisfying

\[ \Delta P = Q\]

Some solutions in 2D:

using PolynomialSolutions
for j in 0:4, i in 0:4
    Q = monomial(i,j)
    P = solve_laplace(Q)
    @show P, Q
end

And some solutions in 3D:

using PolynomialSolutions
for k in 0:4, j in 0:4, i in 0:4
    Q = monomial(i,j,k)
    P = solve_laplace(Q)
    @show P, Q
end

Bi-Laplace

The function solve_bilaplace takes a Polynomial Q and returns a Polynomial P satisfying

\[ \Delta^2 P = Q\]

Some solutions in 2D:

using PolynomialSolutions
for j in 0:4, i in 0:4
    Q = monomial(i,j)
    P = solve_bilaplace(Q)
    @show P, Q
end

And some solutions in 3D:

using PolynomialSolutions
for k in 0:4, j in 0:4, i in 0:4
    Q = monomial(i,j,k)
    P = solve_bilaplace(Q)
    @show P, Q
end

Stokes

The function solve_stokes takes an NTuple of Polynomials Q and returns two Polynomials, U and P, satisfying

\[ \begin{align*} \Delta U - \nabla P &= Q \\ \nabla \cdot U &= 0 \end{align*}\]

using PolynomialSolutions
Q   = (monomial(1,0),monomial(0,0))
U,P = solve_stokes(Q)
@show U,P

Docstrings

PolynomialSolutions.Polynomial — Type

struct Polynomial{N,T}

A polynomial in N variables with coefficients of type T.

PolynomialSolutions.degree — Method

degree(p::Polynomial)

The largest degree of any monomial in p.

PolynomialSolutions.derivative — Method

derivative(p::Polynomial, i::Int)

Differentiate p with respect to the ith variable.

PolynomialSolutions.gradient — Method

gradient(p::Polynomial)

Return an N-tuple of the derivatives of p with respect to each variable.

PolynomialSolutions.multiply_by_r — Method

multiply_by_r(p::Polynomial, k::Int = 2)

Multiply a polynomial p by the monomial r^k, where r = |𝐱| and k is an even positive integer.

PolynomialSolutions.solve_bilaplace — Method

solve_bilaplace(Q::Polynomial)

Compute a polynomial solution to Δ²P = Q.

PolynomialSolutions.solve_elastodynamics — Method

solve_elastodynamics(Q::SVector{N,Polynomial{N,T}};ρ=1,μ=1,ν=1/4,ω=1)

Compute a vector of polynomials U satisfying -μ/(1-2ν) ∇(div U) - μ ΔU - μ k₂² U = Q.

PolynomialSolutions.solve_elastostatic — Method

solve_elastostatic(Q::SVector{N,Polynomial{N,T}};μ=1,ν=1)

Compute a vector of polynomials U satisfying μ/(1-2ν) ∇(div U) + μΔU = Q.

PolynomialSolutions.solve_helmholtz — Method

solve_helmholtz(Q::Polynomial,k=1)

Return the unique polynomial P satisfying ΔP + k²P = Q.

PolynomialSolutions.solve_laplace — Method

solve_laplace(Q::Polynomial)

Return a polynomial P satisfying ΔP = Q.

PolynomialSolutions.solve_maxwell — Method

solve_maxwell(J::SVector{N,Polynomial{N,T}}, ρ::Polynomial{N, T};ϵ=1,μ=1,ω=1)

Compute a pair of vectors of polynomials E and H satisfying the Maxwell system.

PolynomialSolutions.solve_stokes — Method

solve_stokes(Q::SVector{N,Polynomial{N,T}};μ=1)

Compute a vector of polynomials U and a polynomial P satisfying μΔU - ∇P = Q with ∇ ⋅ U = 0.